Fix #2545, migrate TBL, ES, and SB arrays to config list
Utilize the compile-time generated list feature of the config module
to replace hardcoded lists of items in ES, SB, and TBL.  In particular
this applies to the ES mempool sizes, SB mempool sizes, and the
spacecraft and processor ID lists in TBL.

Note that the TBL code was not in accordance with coding standards as
it was (by default) compiled out via an #if directive, and thus not
being tested as it should be.  This fixes that, by removing the
conditional compile and always testing the code.