Fix #???, address sanitizer failure in FS
Add self-defense check in CFE_FS_ParseInputFileNameEx() in case the input
buffer size passed in is greater than the length of the input string.

Technically, this could be argued as an incorrect call/input, but it
is something the function could check for and avoid reading past the
final null char.